requirements.txt
transformers
torch
openai-whisper
yt-dlp
accelerate
